Moved more stuff to the cli module.
